implementacion del formik de forma directa
