+++ .gitignore
@@ -71,4 +71,3 @@ Thumbs.db
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-Package-lock.json
