+++ package.json
@@ -9,7 +9,6 @@
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^5.1.3",
     "formik": "^2.2.9",
-    "node": "^16.13.1",
     "react": "^16.13.1",
     "react-bootstrap": "^2.1.2",
     "react-dom": "^16.13.1",
